Fix: Remove dependency on Assert for GHC < 9.4.X
I tried defining Assert manually, but it seems that the behaviour for
when type errors are reported is different for older GHC versions -
specifically `SatisfiesBound` was reporting a type error because
`BitsNat b <=? BitsNat (BitWidth t)` could not be solved.

An easy solution here is to (on older GHC versions) just use `(<=)`.
This will lead to lower-quality error messages, but will at least work.